backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
drf-nested-routers==0.93.4
django-filter==24.1
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
